Refactor EdgeGrid signer
- The `edgegrid` package is now in a sub-directory containing only the signer
