+++ edgegrid/README.md
@@ -1,0 +1,2 @@
+# Akamai Edgegrid 
+A golang package which facilitates authenticating and signing requests made to [Akamai OPEN APIs](https://developer.akamai.com).
